+++ src/bot/commands/match_commands.py
@@ -8,7 +8,10 @@ from src.bot.commands import helpers
 
 def run_last_match_command(update, context):
     chat_id = update.message.chat_id
-    telegram_handle = update.message.from_user.username
+    try:
+        telegram_handle = context.args[0]
+    except(IndexError, ValueError):
+        telegram_handle = update.message.from_user.username
 
     user = user_services.lookup_user_by_telegram_handle(telegram_handle)