+++ CHANGELOG.md
@@ -10,5 +10,8 @@ CHANGELOG
 
 # Features
 
+- Usernames are now stored and retrieved case-insensitive
+
 # Commands
 
+- /rank, /lastmatch and /recents now let you look up someone else's data by passing their username as an argument